bugfix: replace N with two const vars
The bug was hidden in the variable called Dushinsky::N. The
variable was used in the early part of the program as number
of normal modes in the molecule (3N - 5/6) but midway though
the old constructor value of the variable is changed to the
number of active normal modes (active subspace size).
I have replaced the variable with two const ints
n_mol_nms and n_active_nms.